Refactor cleanup to use server patterns and middleware
I must have missed this package when doing the previous refactors. This will improve our observability and logging visibility for the cleanup workers.
