+++ cmd/cleanup-exposure/main.go
@@ -27,7 +27,6 @@ import (
 	"github.com/google/exposure-notifications-server/pkg/logging"
 	_ "github.com/google/exposure-notifications-server/pkg/observability"
 	"github.com/google/exposure-notifications-server/pkg/server"
-	"github.com/gorilla/mux"
 )
 
 func main() {
@@ -64,20 +63,16 @@ func realMain(ctx context.Context) error {
 	}
 	defer env.Close(ctx)
 
-	handler, err := cleanup.NewExposureHandler(&config, env)
+	cleanupExposureServer, err := cleanup.NewExposureServer(&config, env)
 	if err != nil {
-		return fmt.Errorf("cleanup.NewExposureHandler: %w", err)
+		return fmt.Errorf("cleaup.NewExposureServer: %w", err)
 	}
-
-	r := mux.NewRouter()
-	r.Handle("/", handler)
-	r.Handle("/health", server.HandleHealthz(env.Database()))
 
 	srv, err := server.New(config.Port)
 	if err != nil {
 		return fmt.Errorf("server.New: %w", err)
 	}
-	logger.Infof("listening on :%s", config.Port)
+	logger.Info("listening on: ", config.Port)
 
-	return srv.ServeHTTPHandler(ctx, r)
+	return srv.ServeHTTPHandler(ctx, cleanupExposureServer.Routes(ctx))
 }
