+++ internal/cleanup/cleanup_test.go
@@ -15,15 +15,9 @@
 package cleanup
 
 import (
-	"fmt"
-	"net/http"
-	"net/http/httptest"
 	"testing"
 	"time"
 
-	"github.com/google/exposure-notifications-server/internal/project"
-	"github.com/google/exposure-notifications-server/internal/serverenv"
-	"github.com/google/exposure-notifications-server/internal/storage"
 	"github.com/google/exposure-notifications-server/pkg/database"
 )
 
@@ -33,172 +27,6 @@ func TestMain(m *testing.M) {
 	testDatabaseInstance = database.MustTestInstance()
 	defer testDatabaseInstance.MustClose()
 	m.Run()
-}
-
-func TestNewExposureHandler(t *testing.T) {
-	t.Parallel()
-
-	ctx := project.TestContext(t)
-	testDB, _ := testDatabaseInstance.NewDatabase(t)
-
-	testCases := []struct {
-		name string
-		env  *serverenv.ServerEnv
-		err  error
-	}{
-		{
-			name: "nil Database",
-			env:  serverenv.New(ctx),
-			err:  fmt.Errorf("missing database in server environment"),
-		},
-		{
-			name: "Fully Specified",
-			env:  serverenv.New(ctx, serverenv.WithDatabase(testDB)),
-			err:  nil,
-		},
-	}
-
-	for _, tc := range testCases {
-		tc := tc
-
-		t.Run(tc.name, func(t *testing.T) {
-			t.Parallel()
-
-			got, err := NewExposureHandler(&Config{}, tc.env)
-			if tc.err != nil {
-				if err.Error() != tc.err.Error() {
-					t.Fatalf("got '%+v': want '%v'", err, tc.err)
-				}
-			} else if err != nil {
-				t.Fatalf("got unexpected error: %v", err)
-			} else {
-				handler, ok := got.(*exposureCleanupHandler)
-				if !ok {
-					t.Fatalf("exposureCleanupHandler does not satisfy http.Handler interface")
-				} else if handler.env != tc.env {
-					t.Fatalf("got %+v: want %v", handler.env, tc.env)
-				}
-			}
-		})
-	}
-}
-
-func TestExposureHandler_ServeHTTP(t *testing.T) {
-	t.Parallel()
-
-	ctx := project.TestContext(t)
-	testDB, _ := testDatabaseInstance.NewDatabase(t)
-	env := serverenv.New(ctx, serverenv.WithDatabase(testDB))
-
-	r, err := http.NewRequestWithContext(ctx, http.MethodGet, "/", nil)
-	if err != nil {
-		t.Fatal(err)
-	}
-	w := httptest.NewRecorder()
-
-	handler, err := NewExposureHandler(&Config{
-		Timeout: 5 * time.Second,
-		TTL:     336 * time.Hour,
-	}, env)
-	if err != nil {
-		t.Fatal(err)
-	}
-	handler.ServeHTTP(w, r)
-
-	if got, want := w.Code, http.StatusOK; got != want {
-		t.Errorf("expected %d to be %d: %s", got, want, w.Body.String())
-	}
-}
-
-func TestNewExportHandler(t *testing.T) {
-	t.Parallel()
-
-	ctx := project.TestContext(t)
-	testDB, _ := testDatabaseInstance.NewDatabase(t)
-	noopBlobstore, _ := storage.NewMemory(ctx, &storage.Config{})
-
-	testCases := []struct {
-		name string
-		env  *serverenv.ServerEnv
-		err  error
-	}{
-		{
-			name: "nil_database",
-			env:  serverenv.New(ctx),
-			err:  fmt.Errorf("missing database in server environment"),
-		},
-		{
-			name: "nil_blobstore",
-			env:  serverenv.New(ctx, serverenv.WithDatabase(testDB)),
-			err:  fmt.Errorf("missing blobstore in server environment"),
-		},
-		{
-			name: "fully_specified",
-			env:  serverenv.New(ctx, serverenv.WithBlobStorage(noopBlobstore), serverenv.WithDatabase(testDB)),
-			err:  nil,
-		},
-	}
-
-	for _, tc := range testCases {
-		tc := tc
-
-		t.Run(tc.name, func(t *testing.T) {
-			t.Parallel()
-
-			got, err := NewExportHandler(&Config{}, tc.env)
-			if tc.err != nil {
-				if err.Error() != tc.err.Error() {
-					t.Fatalf("got %+v: want %v", err, tc.err)
-				}
-			} else if err != nil {
-				t.Fatalf("got unexpected error: %v", err)
-			} else {
-				handler, ok := got.(*exportCleanupHandler)
-				if !ok {
-					t.Fatal("exportCleanupHandler does not satisfy http.Handler interface")
-				} else if handler.env != tc.env {
-					t.Fatalf("got %+v: want %v", handler.env, tc.env)
-				}
-			}
-		})
-	}
-}
-
-func TestExportHandler_ServeHTTP(t *testing.T) {
-	t.Parallel()
-
-	ctx := project.TestContext(t)
-	testDB, _ := testDatabaseInstance.NewDatabase(t)
-
-	bs, err := storage.NewMemory(ctx, &storage.Config{
-		Type: "MEMORY",
-	})
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	env := serverenv.New(ctx,
-		serverenv.WithDatabase(testDB),
-		serverenv.WithBlobStorage(bs))
-
-	r, err := http.NewRequestWithContext(ctx, http.MethodGet, "/", nil)
-	if err != nil {
-		t.Fatal(err)
-	}
-	w := httptest.NewRecorder()
-
-	handler, err := NewExportHandler(&Config{
-		Timeout: 5 * time.Second,
-		TTL:     336 * time.Hour,
-	}, env)
-	if err != nil {
-		t.Fatal(err)
-	}
-	handler.ServeHTTP(w, r)
-
-	if got, want := w.Code, http.StatusOK; got != want {
-		t.Errorf("expected %d to be %d: %s", got, want, w.Body.String())
-	}
 }
 
 func TestCutoffDate(t *testing.T) {
